feat: Add PDF metadata toggle functionality
The code changes add functionality to toggle the metadata of a PDF file between "Confidential" and "Not Confidential". This is achieved by adding the `set_metadata()` and `unset_metadata()` functions, which modify the PDF metadata using the PyPDF2 library. The changes also include the implementation of a GUI window using the tkinter library, where the user can select a PDF file and toggle its metadata.